feat: Memoize breadcrumbs
