+++ src/ducks/categories/CategoriesHeader.jsx
@@ -1,4 +1,4 @@
-import React, { Fragment, useCallback } from 'react'
+import React, { Fragment, useCallback, useMemo } from 'react'
 import PropTypes from 'prop-types'
 import cx from 'classnames'
 
@@ -9,6 +9,7 @@ import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 import Switch from 'cozy-ui/transpiled/react/MuiCozyTheme/Switch'
 import Stack from 'cozy-ui/transpiled/react/Stack'
+import { useRouter } from 'components/RouterContext'
 
 import { AccountSwitch } from 'ducks/account'
 import BackButton from 'components/BackButton'
@@ -36,6 +37,27 @@ const stCategory = catStyles['bnk-table-category-category']
 const stPercentage = catStyles['bnk-table-percentage']
 const stTotal = catStyles['bnk-table-total']
 const stTableCategory = catStyles['bnk-table-category']
+
+const makeBreadcrumbs = (router, categoryName, subcategoryName, t) => {
+  const breadcrumbs = [
+    {
+      name: t('Categories.title.general'),
+      onClick: () => router.push('/analysis/categories')
+    }
+  ]
+  if (categoryName) {
+    breadcrumbs.push({
+      name: t(`Data.categories.${categoryName}`),
+      onClick: () => router.push(`/analysis/categories/${categoryName}`)
+    })
+  }
+  if (subcategoryName) {
+    breadcrumbs.push({
+      name: t(`Data.subcategories.${subcategoryName}`)
+    })
+  }
+  return breadcrumbs
+}
 
 const IncomeToggle = ({ withIncome, onToggle }) => {
   const theme = useCozyTheme()
@@ -115,7 +137,6 @@ const CategoriesHeader = props => {
   const { t } = useI18n()
   const { isMobile } = useBreakpoints()
   const {
-    breadcrumbItems,
     emptyIcon,
     hasAccount,
     selectedCategory,
@@ -123,7 +144,9 @@ const CategoriesHeader = props => {
     onWithIncomeToggle,
     categories,
     chartSize,
-    isFetching
+    isFetching,
+    categoryName,
+    subcategoryName
   } = props
 
   const hasData = categories.length > 0 && categories[0].transactionsNumber > 0
@@ -132,6 +155,11 @@ const CategoriesHeader = props => {
   const transactionsTotal = getTransactionsTotal(categories)
   const params = useParams()
   const isSubcategory = onSubcategory(params)
+  const router = useRouter()
+  const breadcrumbItems = useMemo(() => {
+    return makeBreadcrumbs(router, categoryName, subcategoryName, t)
+  }, [router, categoryName, subcategoryName, t])
+
   const accountSwitch = (
     <CategoryAccountSwitch
       selectedCategory={selectedCategory}
@@ -254,7 +282,8 @@ CategoriesHeader.defaultProps = {
 }
 
 CategoriesHeader.propTypes = {
-  breadcrumbItems: PropTypes.array.isRequired,
+  categoryName: PropTypes.string,
+  subcategoryName: PropTypes.string,
   selectedCategory: PropTypes.object,
   withIncome: PropTypes.bool.isRequired,
   onWithIncomeToggle: PropTypes.func.isRequired,
