+++ src/ducks/settings/AccountsSettings.jsx
@@ -1,4 +1,4 @@
-import React, { memo } from 'react'
+import React from 'react'
 
 import groupBy from 'lodash/groupBy'
 import sortBy from 'lodash/sortBy'
